5.1 SR4

+++ Library/PaloBrowser/CubeAnalyzerJob.cpp
@@ -75,7 +75,7 @@ void CubeAnalyzerJob::compute()
 	}
 //	pdims = &dims;
 
-	uint32_t numResult = 0;
+	double numResult = 0;
 	PCubeArea cubeArea;
 	string pathStringMessage = "";
 	string message1;
